requirements.txt
flask>=2.3.3
PyGithub>=2.1.1
requests==2.31.0
python-dotenv==0.19.0